added support for new initialization protocol to etcd client
