+++ config-sample.json
@@ -8,6 +8,7 @@
 
     "state": {
       "strategy": "etcd",
+      "peers": ["10.0.0.1:12345", "10.0.0.2:12345"],
       "endpoints": ["10.0.0.1"],
       "timeout": 2
     }
