+++ examples/configs/etcd.json
@@ -8,6 +8,7 @@
 
     "state": {
       "strategy": "etcd",
+      "peers": ["10.0.0.1:12345", "10.0.0.2:12345"],
       "endpoints": ["172.28.128.7:2379", "172.28.128.8:2379", "172.28.128.9:2379"],
       "timeout": 3
     }
