+++ examples/configs/state-server.json
@@ -8,7 +8,7 @@
 
     "state": {
       "strategy": "state-server",
-      "endpoints": ["127.0.0.1:8080", "127.0.0.1:8081", "127.0.0.1:8082"],
+      "peers": ["10.0.0.1:12345", "10.0.0.2:12345"],
       "timeout": 3
     }
   }
